Disabled the loading and saving of the defaultsettings.mat file for now.  Preliminary fix for the splash screen bug, but won't work until defaultsettings.mat is saved again.
The defaultsettings.mat file can cause incompatibilities if it exists, but has the wrong elements in it.  The GUI doesn't properly identify all possible problems.  For now, always create a new defaultsettings structure on runtime. 

The splash screen bug was caused by the splash screen code overwriting the correct version and clientversion strings before saving the defaultsettings.mat file.  This is fixed, but without re-enabling saving of the file, the code can never remember the setting.